Verified single-threaded edge-based iteration

+++ src/core/utility.hpp
@@ -6,6 +6,7 @@
 namespace CORE
 {
     /// https://stackoverflow.com/questions/27086195/linear-index-upper-triangular-matrix
+    /// WARNING: Watch for potential overflowing
     /// n: width and length
     /// # non-zero elem: n * (n - 1) / 2
     /// k:
@@ -17,15 +18,15 @@ namespace CORE
     /// 3 |  0   0   0   0  a9
     /// 4 |  0   0   0   0   0
 
-    inline int linearize_upper_triangle_matrix_index(int i, int j, int n)
+    inline size_t linearize_upper_triangle_matrix_index(size_t i, size_t j, size_t n)
     {
         return (n * (n - 1) / 2) - (n - i) * ((n - i) - 1) / 2 + j - i - 1;
     }
 
-    inline std::pair<int, int> delinearize_upper_triangle_matrix_index(int k, int n)
+    inline std::pair<size_t, size_t> delinearize_upper_triangle_matrix_index(size_t k, size_t n)
     {
-        int i = n - 2 - static_cast<int>(std::floor(std::sqrt(-8 * k + 4 * n * (n - 1) - 7) / 2.0 - 0.5));
-        int j = k + i + 1 - n * (n - 1) / 2 + (n - i) * ((n - i) - 1) / 2;
+        size_t i = n - 2 - static_cast<size_t>(std::floor(std::sqrt(-8 * k + 4 * n * (n - 1) - 7) / 2.0 - 0.5));
+        size_t j = k + i + 1 - n * (n - 1) / 2 + (n - i) * ((n - i) - 1) / 2;
         return {i, j};
     }
 }
